A different way to select films
Selects films by generating a film id for the title by adding a random six digit number to the end of a tt0 string. This creates these ids until one that has all the necessary fields (none of them are N/A) and is not part of a TV show appear. This greatly limits the number of films availabe to make the game much more winable than just using a random id, however, these are fairly rare compared to invalid or unusable ids, so a lot of requests need to be sent to return one of these useable films which can cause the program to crash. I think it would be better to just store the films in an array like before, I was just trying this out here, probably best leave this out of game in the end.